Use external support packages in CLI
