+++ packages/cli-utils/src/utils/error.ts
@@ -1,0 +1,6 @@
+export class TadaError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = 'TadaError';
+  }
+}
